+++ internal/file/file.go
@@ -1,4 +1,4 @@
-package util
+package file
 
 import (
 	"io"
