+++ pkg/types/types.go
@@ -1,0 +1,10 @@
+package types
+
+import "io"
+
+type Page struct {
+	FilePath string
+	Handler  PageHandler
+}
+
+type PageHandler func(w io.Writer)
